Move ISendZcashAdapter, handle rescan failures
The send adapter interface belongs in core/adapters per convention.
A failing rescan no longer leaves the loading state stuck.
